Extract Smart Race Solver helper components to components/Helpers
